Move actiontyper to sykmeldinger_actions file. Refactor connect() in store.js and combineReducers() in OppfolgingsplanerSide.js
