--- js/actions/sykmeldinger/actiontyper.js
@@ -1,4 +1,0 @@
-export const HENTER_DINE_SYKMELDTE_MED_SYKMELDINGER = 'HENTER_DINE_SYKMELDTE_MED_SYKMELDINGER';
-export const DINE_SYKMELDTE_MED_SYKMELDINGER_HENTET = 'DINE_SYKMELDTE_MED_SYKMELDINGER_HENTET';
-export const HENT_DINE_SYKMELDTE_MED_SYKMELDINGER_FEILET = 'HENT_DINE_SYKMELDTE_MED_SYKMELDINGER_FEILET';
-export const HENT_DINE_SYKMELDTE_MED_SYKMELDINGER_FORESPURT = 'HENT_DINE_SYKMELDTE_MED_SYKMELDINGER_FORESPURT';
