+++ js/reducers/dineSykmeldteMedSykmeldinger.js
@@ -1,4 +1,4 @@
-import * as actiontyper from '../actions/sykmeldinger/actiontyper';
+import * as actiontyper from '../actions/sykmeldinger/sykmeldinger_actions';
 
 const initiellState = {
   henter: false,
